moving some classes around, added ability for gradle to run the unit tests
