+++ src/main/java/rdw/reporting/repository/TranslationRepository.java
@@ -1,6 +1,6 @@
 package rdw.reporting.repository;
 
-import rdw.reporting.models.Translation;
+import rdw.reporting.model.Translation;
 
 import javax.validation.constraints.NotNull;
 import java.util.Locale;
