+++ src/main/java/rdw/reporting/web/endpoints/ReportingController.java
@@ -4,10 +4,9 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 import org.springframework.jdbc.core.JdbcTemplate;
-import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.*;
-import rdw.reporting.models.ICA;
-import rdw.reporting.repositories.InterimComprehensiveQuery;
+import rdw.reporting.model.ICA;
+import rdw.reporting.repository.InterimComprehensiveQuery;
 
 import java.util.*;
