+++ src/test/java/rdw/reporting/services/impl/TranslationServiceImplTest.java
@@ -3,8 +3,8 @@ package rdw.reporting.services.impl;
 import com.google.common.collect.Sets;
 import org.junit.Before;
 import org.junit.Test;
-import rdw.reporting.models.ImmutableTranslation;
-import rdw.reporting.models.Translation;
+import rdw.reporting.model.ImmutableTranslation;
+import rdw.reporting.model.Translation;
 import rdw.reporting.repository.TranslationRepository;
 import rdw.reporting.service.TranslationService;
 import rdw.reporting.service.TranslationServiceImpl;
